Changed some values from int32 to int64

+++ model/model.go
@@ -5,31 +5,31 @@ type RspamdStatistics struct {
 	Scanned  int32 `json:"scanned"`
 	Learned  int32 `json:"learned"`
 	Actions struct {
-		Reject         int32 `json:"reject"`
-		SoftReject     int32 `json:"soft reject"`
-		RewriteSubject int32 `json:"rewrite subject"`
-		AddHeader      int32 `json:"add header"`
-		Greylist       int32 `json:"greylist"`
-		NoAction       int32 `json:"no action"`
+		Reject         int64 `json:"reject"`
+		SoftReject     int64 `json:"soft reject"`
+		RewriteSubject int64 `json:"rewrite subject"`
+		AddHeader      int64 `json:"add header"`
+		Greylist       int64 `json:"greylist"`
+		NoAction       int64 `json:"no action"`
 	} `json:"actions"`
-	SpamCount             int32 `json:"spam_count"`
-	HamCount              int32 `json:"ham_count"`
-	Connections           int32 `json:"connections"`
-	ControlConnections    int32 `json:"control_connections"`
-	PoolsAllocated        int32 `json:"pools_allocated"`
-	PoolsFreed            int32 `json:"pools_freed"`
-	BytesAllocated        int32 `json:"bytes_allocated"`
-	ChunksAllocated       int32 `json:"chunks_allocated"`
-	SharedChunksAllocated int32 `json:"shared_chunks_allocated"`
-	ChunksFreed           int32 `json:"chunks_freed"`
-	ChunksOversized       int32 `json:"chunks_oversized"`
-	Fragmented            int32 `json:"fragmented"`
-	TotalLearns           int32 `json:"total_learns"`
+	SpamCount             int64 `json:"spam_count"`
+	HamCount              int64 `json:"ham_count"`
+	Connections           int64 `json:"connections"`
+	ControlConnections    int64 `json:"control_connections"`
+	PoolsAllocated        int64 `json:"pools_allocated"`
+	PoolsFreed            int64 `json:"pools_freed"`
+	BytesAllocated        int64 `json:"bytes_allocated"`
+	ChunksAllocated       int64 `json:"chunks_allocated"`
+	SharedChunksAllocated int64 `json:"shared_chunks_allocated"`
+	ChunksFreed           int64 `json:"chunks_freed"`
+	ChunksOversized       int64 `json:"chunks_oversized"`
+	Fragmented            int64 `json:"fragmented"`
+	TotalLearns           int64 `json:"total_learns"`
 	Statfiles []struct {
 		Revision  int32  `json:"revision"`
 		Used      int32  `json:"used"`
 		Total     int32  `json:"total"`
-		Size      int32  `json:"size"`
+		Size      int64  `json:"size"`
 		Symbol    string `json:"symbol"`
 		Type      string `json:"type"`
 		Languages int32  `json:"languages"`
